pyproject.toml
[project]
name = "gemini-rag-example"
version = "0.1.0"
description = "Hello world RAG application using LangChain, ChromaDB, and Google Gemini."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=1.0.4",
    "chromadb>=0.5.5",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
rag-demo = "rag_demo.cli:main"

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
